requirements.txt
# المكتبات التي يحتاجون لتثبيتها
mistralai
python-dotenv
